Green Box Fix
Fix for the green box bug that occurs with durability bars. There is a
bug where UPDATE_INVENTORY_DURABILITY is not called when an item being
equipped on the paper doll has the same transmog as the item it is
replacing.

I have opted to use PLAYER_EQUIPMENT_CHANGED instead.

I didn't get to fully test as the servers went down. We can always
hookfunc the paperdoll frame opening to call PLAYER_EQUIPMENT_CHANGED.