For April the 1-st
In patchnotes somethingsimilar to "You may be aware that during
preparation for Battle for Azeroth Blizzard plans to do a stat and item
level squish. Our sources reveal that a division by 4.01 will occur.
That's why we released this DCS version to combat it. We also provided a
way through Interface Options to see your real stats in case you want
it."
Setting of date is  in DCSLayouts.lua (280-281 and 342-343). Setting of
multiplier DCSTables.lua (line 246).